web/requirements.txt
anyio==3.6.2
attrs==22.2.0
Authlib==1.2.0
autopep8==2.0.2
aws-cdk-lib==2.72.0
aws-cdk.asset-awscli-v1==2.2.123
aws-cdk.asset-kubectl-v20==2.1.1
aws-cdk.asset-node-proxy-agent-v5==2.0.100
boto3==1.26.102
botocore==1.29.103
cattrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
constructs==10.1.296
cryptography==40.0.1
Deprecated==1.2.13
exceptiongroup==1.1.1
fastapi==0.95.0
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-resources==5.12.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsii==1.79.0
Markdown==3.4.3
MarkupSafe==2.1.2
packaging==23.1
pluggy==1.0.0
publication==0.0.3
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.7
PyGithub==1.58.1
PyJWT==2.6.0
PyNaCl==1.5.0
pytest==7.3.1
pytest-asyncio==0.21.0
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
tomli==2.0.1
typeguard==2.13.3
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
wrapt==1.15.0
